Fix build error if output directory does not already exist.

+++ src/components/publisher.ts
@@ -20,10 +20,9 @@ const publish = (graph: Graph) => {
     log.msg(`Publishing to output directory`);
     log.msg(`> Output dir: ${fullOutPath}`);
     log.msg(`> ${graph.size()} files to publish`);
-    // TODO: Make this optional.
-    fs.rmSync(fullOutPath, {recursive: true});
-
-
+    if (fs.existsSync(fullOutPath)) {
+        fs.rmSync(fullOutPath, {recursive: true});
+    }
 
     graph.visit(publishItem);
 }
